refactor(desktop): retire http-fence parser for the lezer grammar
